requirements.txt
numpy
torch
rasterio
pystac-client
joblib
pandas
Pillow
config
planetary-computer
requests
